admin/requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
pika==1.1.0